+++ src/main/java/com/aetherteam/aether/inventory/menu/AetherAccessoriesMenu.java
@@ -1,9 +1,9 @@
 package com.aetherteam.aether.inventory.menu;
 
 import com.aetherteam.aether.accessories.api.AccessoriesAPI;
-import com.aetherteam.aether.accessories.api.menu.AccessoriesBasedSlot;
 import com.aetherteam.aether.accessories.api.slot.SlotTypeReference;
 import com.aetherteam.aether.inventory.AetherAccessorySlots;
+import com.aetherteam.aether.inventory.menu.slot.AccessoriesBasedSlot;
 import com.aetherteam.aether.mixin.mixins.common.accessor.AbstractContainerMenuAccessor;
 import com.aetherteam.aether.mixin.mixins.common.accessor.CraftingMenuAccessor;
 import net.minecraft.resources.Identifier;